motor cancel fixed
