+++ app/js/config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  baseUrl: 'http://localhost:5000'
+};
