+++ app/js/entry.js
@@ -1,0 +1,5 @@
+const angular = require('angular');
+const bikeApp = angular.module('bikeApp', []);
+
+require('./pedals')(bikeApp);
+require('./motors')(bikeApp);
