+++ app/js/lib/handle-error.js
@@ -1,0 +1,4 @@
+module.exports = function(error) {
+  console.log(error);
+  this.errors = (this.errors || []).push(error);
+};
