+++ app/js/lib/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  handleError: require('./handle-error')
+};
