+++ app/js/motors/controllers/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./motor-controller')(app);
+};
