+++ app/js/motors/index.js
@@ -1,0 +1,4 @@
+module.exports = function(app) {
+  require('./controllers')(app);
+  // require('./directives')(app);
+};
