+++ app/js/pedals/controllers/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./pedal-controller')(app);
+};
