+++ app/js/pedals/directives/index.js
@@ -1,0 +1,4 @@
+module.exports = function(app) {
+  require('./pedal-form')(app);
+  require('./pedal-list-item')(app);
+};
