+++ app/js/pedals/index.js
@@ -1,0 +1,4 @@
+module.exports = function(app) {
+  // require('./directives')(app);
+  require('./controllers')(app);
+};
